Store both state and machine in app-db
This uses a different integration with clj-statecharts, one which
expects the machine and state to be stored in the app-db. In this way,
the re-frame acts like what clj-statecharts calls a "service". That is,
it's a stateful wrapper around the immutable machines and states.

The value of doing it this way is mostly in the new integration. This
integration registers only a few event handlers. Those same event
handlers can transition any machine and state, though they have to be
told which machine and state to act upon.

This lets us manage the memory consumed by the machines and the states,
and avoids accumulating handlers in the global event handler registry.