Fix callbacks on embedded machines
When using an embedded machine, you can choose to transition to a custom
location in the state machine on failure or success (for example, to
implement polling). But, we were depending on reaching the ::halted and
::loaded states to dispatch the request's :on-failure and :on-success.

This commit fixes that, so that the request callbacks are always called,
even if you advance to a different part of the state machine.

It also ensures that the :on-error callback is called after every failed
xhrio, not just the first one.